+++ backend/services/rating-service/rating/src/main/java/vn/uit/edu/msshop/rating/domain/model/valueobject/RatingPoint.java
@@ -3,9 +3,9 @@ package vn.uit.edu.msshop.rating.domain.model.valueobject;
 import vn.edu.uit.msshop.shared.domain.exception.DomainException;
 
 public record RatingPoint(
-        byte value) {
-    public static final byte MIN_VALUE = 1;
-    public static final byte MAX_VALUE = 5;
+        int value) {
+    public static final int MIN_VALUE = 1;
+    public static final int MAX_VALUE = 5;
 
     public RatingPoint {
         if (value < MIN_VALUE) {
